+++ server.js
@@ -2,7 +2,7 @@ const express = require('express');
 const bodyParser = require('body-parser');
 const cors = require('cors');
 const mongoose = require('mongoose');
-const passport = require('./Controllers/passport');
+const passport = require('./Service/UserAuthenticate');
 const app = express();
 const PORT = 3000;
 const routes = require('./Routes');
